+++ source_gen/source_gen/test/constants_test.dart
@@ -68,29 +68,34 @@ void main() {
     test('should read a String', () {
       expect(constants[0].isString, isTrue);
       expect(constants[0].stringValue, 'Hello');
+      expect(constants[0].isAny, isTrue);
       expect(constants[0].anyValue, 'Hello');
     });
 
     test('should read an Int', () {
       expect(constants[1].isInt, isTrue);
       expect(constants[1].intValue, 1234);
+      expect(constants[1].isAny, isTrue);
       expect(constants[1].anyValue, 1234);
     });
 
     test('should read a Bool', () {
       expect(constants[2].isBool, isTrue);
       expect(constants[2].boolValue, isTrue);
+      expect(constants[2].isAny, isTrue);
       expect(constants[2].anyValue, isTrue);
     });
 
     test('should read a Null', () {
       expect(constants[3].isNull, isTrue);
+      expect(constants[3].isAny, isTrue);
       expect(constants[3].anyValue, isNull);
     });
 
     test('should read an arbitrary object', () {
       final constant = constants[4];
 
+      expect(constant.isAny, isFalse);
       expect(() => constant.anyValue, throwsFormatException);
 
       expect(constant.read('aString').stringValue, 'Hello');
@@ -113,6 +118,7 @@ void main() {
       expect(constants[6].isList, isTrue, reason: '${constants[6]}');
       expect(constants[6].listValue.map((c) => new ConstantReader(c).intValue),
           [1, 2, 3]);
+      expect(constants[6].isAny, isFalse);
       expect(() => constants[6].anyValue, throwsFormatException);
     });
 
@@ -123,17 +129,20 @@ void main() {
               key: (k, _) => new ConstantReader(k).intValue,
               value: (_, v) => new ConstantReader(v).stringValue),
           {1: 'A', 2: 'B'});
+      expect(constants[7].isAny, isFalse);
       expect(() => constants[7].anyValue, throwsFormatException);
     });
 
     test('should read a double', () {
       expect(constants[9].isDouble, isTrue);
       expect(constants[9].doubleValue, 1.23);
+      expect(constants[9].isAny, isTrue);
       expect(constants[9].anyValue, 1.23);
     });
 
     test('should read a Symbol', () {
       expect(constants[10].isSymbol, isTrue);
+      expect(constants[10].isAny, isTrue);
       expect(constants[10].symbolValue, #shanna);
       expect(constants[10].anyValue, #shanna);
     });
@@ -141,6 +150,7 @@ void main() {
     test('should read a Type', () {
       expect(constants[11].isType, isTrue);
       expect(constants[11].typeValue.name, 'DateTime');
+      expect(constants[11].isAny, isFalse);
       expect(() => constants[11].anyValue, throwsFormatException);
     });