+++ source_gen/source_gen/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.7.1
+
+### ConstantReader
+
+* Added `isAny`.
+* Allow `anyValue` to return `null` if the value is actually `null`.
+* Fixed a bug where a `Symbol` would return as a `String` for `anyValue`.
+
 ## 0.7.0
 
 * **Breaking changes**: See [the wiki](https://github.com/dart-lang/source_gen/wiki/Migrating-to-0.7.x)